1. Created BasePage 2. Merged 3 RefreshButton feature and Step Definitions files into one 3. Merged 2 ExportGridDropdown feature and Step Definitions files into one 4. Updated CukesRunner file with plugins and deleted JSON Runner 5. Cleaned up VehiclesPage file by removing duplicate WebElements
