+++ src/test/java/com/cydeo/runner/CukesRunner.java
@@ -4,8 +4,11 @@ import io.cucumber.junit.CucumberOptions;
 import org.junit.runner.RunWith;
     @RunWith(Cucumber.class)
     @CucumberOptions(
+
+            plugin = {"html:target/cucumber_report.html","json:target/cucumber-json-report.json"},
             features = "src/test/resources/features" ,
             glue = "com/cydeo/step_definitions",
             dryRun=false,
             tags = "@Test3"
+
     ) public class CukesRunner {}
