+++ src/test/java/com/cydeo/step_definitions/ExportGridDropdown_StepDefinitions.java
@@ -1,5 +1,6 @@
 package com.cydeo.step_definitions;
 
+import com.cydeo.pages.BasePage;
 import com.cydeo.pages.VehiclesPage;
 import com.cydeo.utilities.Driver;
 import com.cydeo.utilities.VyTrackUtils;
@@ -12,24 +13,53 @@ import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
-public class ExportGritDropdownLocation_StepDef {
+public class ExportGridDropdown_StepDefinitions {
 
     VehiclesPage vehiclesPage = new VehiclesPage();
     Actions actions;
-    WebDriverWait wait = new WebDriverWait(Driver.getDriver(),30);
+    WebDriverWait wait = new WebDriverWait(Driver.getDriver(), 30);
+
+    BasePage basePage = new BasePage();
+
+    @Given("^user \"([^\"]*)\" \"([^\"]*)\" is on the vehicle page$")
+    public void user_is_on_the_vehicle_page(String arg1, String arg2) {
+        VyTrackUtils.login(arg1, arg2);
+        VyTrackUtils.goToVehiclesPage();
+    }
+
+    @When("^user clicks on Export Grid dropdown button$")
+    public void user_clicks_on_export_grid_dropdown_button() {
+        VehiclesPage vehiclesPage = new VehiclesPage();
+        WebDriverWait wait = new WebDriverWait(Driver.getDriver(), 30);
+        wait.until(ExpectedConditions.elementToBeClickable(vehiclesPage.exportGridDropdownButton));
+        vehiclesPage.exportGridDropdownButton.click();
+    }
+
+    @Then("^user should see two options: CSV and XLSX$")
+    public void user_should_see_two_options_csv_and_xlsx() {
+        VehiclesPage vehiclesPage = new VehiclesPage();
+        Assert.assertTrue(vehiclesPage.csvButton.isDisplayed());
+        Assert.assertTrue(vehiclesPage.xlsxButton.isDisplayed());
+        VyTrackUtils.vyTrack_logout();
+    }
+
+    /**
+     * Ganjina's Stuff:
+     */
 
     @Given("user {string} {string} is on the All cars page")
     public void userIsOnTheAllCarsPage(String username, String password) {
         VyTrackUtils.login(username, password);
         VyTrackUtils.goToVehiclesPage();
-
     }
+
     @When("user moves to Export Grid Dropdown button")
     public void user_moves_to_export_grid_dropdown_button() {
         wait.until(ExpectedConditions.elementToBeClickable(vehiclesPage.exportGridDropdownButton));
         actions = new Actions(Driver.getDriver());
         actions.moveToElement(vehiclesPage.exportGridDropdownButton).pause(3000).perform();
     }
+
     @Then("user sees the Export Grid Dropdown button on the left side of the page")
     public void user_sees_the_export_grid_dropdown_button_on_the_left_side_of_the_page() {
         Point location = vehiclesPage.exportGridDropdownButton.getLocation();
@@ -40,11 +70,11 @@ public class ExportGritDropdownLocation_StepDef {
         int height = vehiclesPage.exportGridDropdownButton.getSize().getHeight();
         int width = vehiclesPage.exportGridDropdownButton.getSize().getWidth();
 
-        Assert.assertTrue("Test failed",((xPos+width) <= winWidth/2) && (yPos + height) <= winHeight/2);
+        Assert.assertTrue("Test failed", ((xPos + width) <= winWidth / 2) && (yPos + height) <= winHeight / 2);
 
-        vehiclesPage.logout();
-
+        basePage.logout();
     }
-
-
 }
+
+//solution for feature 1
+
